webkit/data/layout_tests/chrome/http/tests/security/xss-DENIED-assign-location-href-javascript-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL http://localhost:8081/chrome/http/tests/security/resources/localPage.html from frame with URL http://127.0.0.1:8081/chrome/http/tests/security/xss-DENIED-assign-location-href-javascript.html. Domains, protocols and ports must match.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
A local page
